% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZOIP.R
\name{dZOIP}
\alias{dZOIP}
\title{Distribucion ZOIP}
\usage{
dZOIP(x, mu = 0.5, sigma = 0.1, p0 = 0.08333333, p1 = 0.08333333,
  family = "R-S", log = FALSE)
}
\arguments{
\item{x, q}{vector de cuantiles}

\item{mu}{vector de parametros de localizacion}

\item{sigma}{vector de parametros de escala}

\item{p0}{parametro de proporcion de ceros}

\item{p1}{parametro de proporcion de unos}

\item{family}{eleccion de la parametrizacion o distribucion deseada}

\item{log, log.p}{logico; si TRUE, las probabilidades de p estaran dadas como log(p).}

\item{lower.tail}{logico; si TRUE (default), probabilidades estaran P[X <= x], en otro caso, P[X > x]}

\item{p}{vector de probabilidades.}

\item{n}{numero de observaciones. si length(n) > 1, el tamañano se tomara como el numero requerido}
}
\description{
La funcion ZOIP() define las distribuciones para datos proporcionales inflados en ceros
y unos mas conocidas, como la distribucion beta y simplex, ademas permite que se tengan
distintas parametrizaciones.La funcion ZOIP() posee 4 parametros dos de ellos es la
porporcion de ceros y unos que quiere que posea la distribucion para datos proporcionales
de su eleccion.
}
\details{
Las funciones dZOIP, pZOIP, qZOIP y rZOIP define la densidad, la funcion de distribucion,
la funcion cuantil y la funcion generadora de numeros aleatorios para la distribucion ZOIP.
}
\examples{
dZOIP(x=0.5, mu = 0.2, sigma = 0.5, p0 = 0.2, p1 = 0.2,family='R-S',log = FALSE)
dZOIP(x=0.5, mu = 0.2, sigma = 3, p0 = 0.2, p1 = 0.2,family='F-C',log = FALSE)
dZOIP(x=0.5, mu = 0.6, sigma = 2.4, p0 = 0.2, p1 = 0.2,family='Original',log = FALSE)
dZOIP(x=0.5, mu = 0.2, sigma = 3, p0 = 0.2, p1 = 0.2,family='Simplex',log = FALSE)
pZOIP(q=0.5, mu = 0.2, sigma = 0.5, p0 = 0.2, p1 = 0.2,family='R-S',log = FALSE)
pZOIP(q=0.5, mu = 0.2, sigma = 3, p0 = 0.2, p1 = 0.2,family='F-C',log = FALSE)
pZOIP(q=0.5, mu = 0.6, sigma = 2.4, p0 = 0.2, p1 = 0.2,family='Original',log = FALSE)
pZOIP(q=0.5, mu = 0.2, sigma = 3, p0 = 0.2, p1 = 0.2,family='Simplex',log = FALSE)
qZOIP(p=0.7, mu = 0.2, sigma = 0.5, p0 = 0.2, p1 = 0.2,family='R-S',log = FALSE)
qZOIP(p=0.7, mu = 0.2, sigma = 3, p0 = 0.2, p1 = 0.2,family='F-C',log = FALSE)
qZOIP(p=0.7, mu = 0.6, sigma = 2.4, p0 = 0.2, p1 = 0.2,family='Original',log = FALSE)
qZOIP(p=0.7, mu = 0.2, sigma = 3, p0 = 0.2, p1 = 0.2,family='Simplex',log = FALSE)
a1<-rZOIP(n=1000, mu = 0.2, sigma = 0.5, p0 = 0.2, p1 = 0.2,family='R-S')
a2<-rZOIP(n=1000, mu = 0.2, sigma = 3, p0 = 0.2, p1 = 0.2,family='F-C')
a3<-rZOIP(n=1000, mu = 0.6, sigma = 2.4, p0 = 0.2, p1 = 0.2,family='Original')
system.time(a4<-rZOIP(n=10, mu = 0.2, sigma = 3, p0 = 0.2, p1 = 0.2,family='Simplex'))

plot(density(a1))
plot(density(a2))
plot(density(a3))
plot(density(a4))
}

