% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pZOIP.R
\name{pZOIP}
\alias{pZOIP}
\title{Distribucion ZOIP}
\usage{
pZOIP(q, mu = 0.5, sigma = 0.1, p0 = 0.08333333, p1 = 0.08333333,
  family = "R-S", lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{vector de cuantiles}

\item{mu}{vector de parametros de localizacion}

\item{sigma}{vector de parametros de escala}

\item{p0}{parametro de proporcion de ceros}

\item{p1}{parametro de proporcion de unos}

\item{family}{eleccion de la parametrizacion o distribucion deseada}

\item{lower.tail}{logico; si TRUE (default), probabilidades estaran P[X <= x], en otro caso, P[X > x]}

\item{log.p}{logico; si TRUE, las probabilidades de p estaran dadas como log(p).}
}
\description{
La funcion pZOIP define la funcion de distribucion acumulada de la distribucion ZOIP.
}
\examples{
pZOIP(q=0.5, mu = 0.2, sigma = 0.5, p0 = 0.2, p1 = 0.2,family='R-S',log = FALSE)
pZOIP(q=0.5, mu = 0.2, sigma = 3, p0 = 0.2, p1 = 0.2,family='F-C',log = FALSE)
pZOIP(q=0.5, mu = 0.6, sigma = 2.4, p0 = 0.2, p1 = 0.2,family='Original',log = FALSE)
pZOIP(q=0.5, mu = 0.2, sigma = 3, p0 = 0.2, p1 = 0.2,family='Simplex',log = FALSE)
}

