% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qZOIP.R
\name{qZOIP}
\alias{qZOIP}
\title{Distribucion ZOIP}
\usage{
qZOIP(p, mu = 0.5, sigma = 0.1, p0 = 0.08333333, p1 = 0.08333333,
  family = "R-S", lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{vector de probabilidades.}

\item{mu}{vector de parametros de localizacion}

\item{sigma}{vector de parametros de escala}

\item{p0}{parametro de proporcion de ceros}

\item{p1}{parametro de proporcion de unos}

\item{family}{eleccion de la parametrizacion o distribucion deseada}

\item{log}{logico; si TRUE, las probabilidades de p estaran dadas como log(p).}
}
\description{
La funcion qZOIP define la funcion cuantil de la distribucion ZOIP.
}
\examples{
qZOIP(p=0.7, mu = 0.2, sigma = 0.5, p0 = 0.2, p1 = 0.2,family='R-S',log = FALSE)
qZOIP(p=0.7, mu = 0.2, sigma = 3, p0 = 0.2, p1 = 0.2,family='F-C',log = FALSE)
qZOIP(p=0.7, mu = 0.6, sigma = 2.4, p0 = 0.2, p1 = 0.2,family='Original',log = FALSE)
qZOIP(p=0.7, mu = 0.2, sigma = 3, p0 = 0.2, p1 = 0.2,family='Simplex',log = FALSE)
}

