% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMMZOIP.R
\name{RMM.ZOIP}
\alias{RMM.ZOIP}
\title{Modelo de Regresion mixto ZOIP}
\usage{
RMM.ZOIP(formula.mu,formula.sigma=~1,formula.p0=~1,formula.p1=~1,data,formula.random,link=c('identity','identity','identity','identity'),family='R-S',optimizer='nlminb',n.points=11,pruning=TRUE)
}
\arguments{
\item{formula.mu}{Formula que define la funcion de regresion para mu, p.e y~x1+x2, es necesario definir la variable respuesta.}

\item{formula.sigma}{Formula que define la funcion de regresion para sigma, p.e ~x1, es necesario definir la variable respuesta.}

\item{formula.p0}{Formula que define la funcion de regresion para p0, p.e ~x1, es necesario definir la variable respuesta.}

\item{formula.p1}{Formula que define la funcion de regresion para p1, p.e ~x1, es necesario definir la variable respuesta.}

\item{data}{Datos en fomato data.frame donde debe contener las nombres de las columnas tal cual como estan en las formulas.}

\item{formula.random}{Formula que define el intercepto aleatorio a tener encuenta en la funcion, p.e formula.random= ~ 1 | X1.}

\item{link}{Eleccion de la funcion de enlace, se elige a cada parametro a estimar de acuerdo a las opciones de familia y formula, ver detalles y vignettes.}

\item{family}{Eleccion de la parametrizacion o distribucion deseada, family='R-S' parametrizacion distribucion beta Rigby y Stasinopoulos, 'F-C' distribucion Beta parametrizacion Ferrari y Cribari-Neto, Original distribucion beta parametrizacion original, 'Simplex' distribucion simplex. por defecto 'R-S'}

\item{optimizer}{Eleccion del optimizador, utilizado para la convergencia de la maxima verosimilitud.}

\item{n.points}{numero de puntos a utilizar en la aproximacion por medio de la cuadratura de gauss-hermite adpatativa. por defecto es 11.}

\item{pruning}{es un valor booleano que indica si se utilizara pruning o no. por defecto es TRUE.}
}
\description{
la funcion RMM.ZOIP ajusta un modelo de regresion mixto ZOIP con interceptos aletorios
normales en el parametro de media y dispersion, la estimación se realiza via maxima verosimilitud
y la cuadratura de gauss-hermite adaptativa con o sin pruning. el modelo puede incluir o no efectos
fijos en cualquiera de sus parametros, asi como puede ser un modelo inflado bilateralmente, unilateralmente o sin parametros
inflados.
}
\examples{

library(ZOIP)
N<-21

Times <- c(2, 10, 20, 40) # cantidad de dias

subject <- rep(1:N, each=length(Times)) # numero de sujetos en la muestra repetidos tantas veces haya dias

Days <- rep(Times, times=N)
b0i <- rep(rnorm(n=N,sd=1), each=length(Times))
b1i <- rep(rnorm(n=N,sd=0.5), each=length(Times))

neta <- (1.6+b0i)-1.3*log(Days)
neta2<-(0.1+b1i)-0.8*log(Days)


mu <- 1 / (1 + exp(-neta))
sigma <- 1 / (1 + exp(-neta2))

p0 <- 0.1
p1 <- 0.1

mu[mu==1] <- 0.999
mu[mu==0] <- 0.001

sigma[sigma==1] <- 0.999
sigma[sigma==0] <- 0.001


family<-'R-S'

Y <- rZOIP(n=length(mu), mu = mu, sigma = sigma ,p0=p0,p1=p1,family=family)
base<-data.frame(Y,Days,subject)


formula.mu <- Y ~ log(Days)
formula.sigma<-~log(Days)
formula.p0<-~1
formula.p1<-~1

formula.random=~1 | subject

link<-c('logit','logit','identity','identity')


optimizer<-'nlminb'
n.points <-11
pruning <- TRUE

mod<-RMM.ZOIP(formula.mu=formula.mu,formula.sigma=formula.sigma,formula.p0=formula.p0,formula.p1=formula.p1,data=base,
              formula.random=formula.random,link=link,family=family,optimizer=optimizer,
              n.points=n.points,pruning=pruning)
mod
}

