% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rZOIP.R
\name{rZOIP}
\alias{rZOIP}
\title{Distribucion ZOIP}
\usage{
rZOIP(n, mu = 0.5, sigma = 0.1, p0 = 0.08333333, p1 = 0.08333333,
  family = "R-S")
}
\arguments{
\item{n}{numero de observaciones. si length(n) > 1, el tamañano se tomara como el numero requerido}

\item{mu}{vector de parametros de localizacion}

\item{sigma}{vector de parametros de escala}

\item{p0}{parametro de proporcion de ceros}

\item{p1}{parametro de proporcion de unos}

\item{family}{eleccion de la parametrizacion o distribucion deseada}
}
\description{
la funcion rZOIP define la funcion generadora de numeros aleatorios para la distribucion ZOIP.
}
\examples{
a1<-rZOIP(n=1000, mu = 0.2, sigma = 0.5, p0 = 0.2, p1 = 0.2,family='R-S')
a2<-rZOIP(n=1000, mu = 0.2, sigma = 3, p0 = 0.2, p1 = 0.2,family='F-C')
a3<-rZOIP(n=1000, mu = 0.6, sigma = 2.4, p0 = 0.2, p1 = 0.2,family='Original')
system.time(a4<-rZOIP(n=10, mu = 0.2, sigma = 3, p0 = 0.2, p1 = 0.2,family='Simplex'))
plot(density(a1))
plot(density(a2))
plot(density(a3))
plot(density(a4))
}

