% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dZOIP.R
\name{dZOIP}
\alias{dZOIP}
\title{Distribucion ZOIP}
\usage{
dZOIP(x, mu = 0.5, sigma = 0.1, p0 = 0.08333333, p1 = 0.08333333,
  family = "R-S", log = FALSE)
}
\arguments{
\item{x}{vector de cuantiles}

\item{mu}{vector de parametros de localizacion}

\item{sigma}{vector de parametros de escala}

\item{p0}{parametro de proporcion de ceros}

\item{p1}{parametro de proporcion de unos}

\item{family}{eleccion de la parametrizacion o distribucion deseada}

\item{log}{logico; si TRUE, las probabilidades de p estaran dadas como log(p).}
}
\description{
La funcion dZOIP define la funcion de densidad de probabilidad de la distribucion ZOIP.
}
\examples{
dZOIP(x=0.5, mu = 0.2, sigma = 0.5, p0 = 0.2, p1 = 0.2,family='R-S',log = FALSE)
dZOIP(x=0.5, mu = 0.2, sigma = 3, p0 = 0.2, p1 = 0.2,family='F-C',log = FALSE)
dZOIP(x=0.5, mu = 0.6, sigma = 2.4, p0 = 0.2, p1 = 0.2,family='Original',log = FALSE)
dZOIP(x=0.5, mu = 0.2, sigma = 3, p0 = 0.2, p1 = 0.2,family='Simplex',log = FALSE)

}

