% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMMZOIP.R
\name{RMM.ZOIP}
\alias{RMM.ZOIP}
\title{ZOIP regression with mixed efects}
\usage{
RMM.ZOIP(formula.mu,formula.sigma=~1,formula.p0=~1,formula.p1=~1,data,formula.random,link=c('identity','identity','identity','identity'),family='R-S',optimizer='nlminb',n.points=11,pruning=TRUE)
}
\arguments{
\item{formula.mu}{Formula que define la funcion de regresion para mu, p.e y~x1+x2, es necesario definir la variable respuesta.}

\item{formula.sigma}{Formula que define la funcion de regresion para el parametro sigma, un valor posible es ~x1, por defecto ~1.}

\item{formula.p0}{Formula que define la funcion de regresion para p0, un valor posible es ~x1, por defecto ~1.}

\item{formula.p1}{Formula que define la funcion de regresion para p1, un valor posible es ~x1, por defecto ~1.}

\item{data}{Es el conjunto de datos en formato data.frame donde debe contener las nombres de las columnas tal cual como estan en las formulas.}

\item{formula.random}{Formula que define el efecto mixto dentro del modelo, debe ser solo el intercepto aleatorio que se tendra en cuenta en el parametro de la media y la dispersion, la estructura admisible es la siguiente formula.random = ~1 | G1, donde G1 es la variable que indica los grupos o sujetos en el modelo, siempre debe ser definido.}

\item{link}{Es un vector con las funciones enlace adecuadas para cada parametro a estimar de acuerdo a las opciones escogidas en los parametros de familia y formula. Si el modelo de regresion no posee covariables se debe utilizar como funcion enlace la opcion identity, independientemente del valor escogido en familia, opciones posibles son logit, log, por defecto link=c('identity','identity','identity','identity').}

\item{family}{Eleccion de la parametrizacion o distribucion deseada, family='R-S' parametrizacion distribucion beta Rigby y Stasinopoulos, 'F-C' distribucion Beta parametrizacion Ferrari y Cribari-Neto, Original distribucion beta parametrizacion original, 'Simplex' distribucion simplex. por defecto 'R-S'.}

\item{optimizer}{Eleccion del optimizador, utilizado para encontrar la convergencia de la maxima verosimilitud. se puede elegir el valor de 'nlminb' o 'optim', por defecto 'nlminb'.}

\item{n.points}{Numero de puntos a utilizar en la aproximacion de la funcion de verosimilitud por medio de la cuadratura de Gauss-Hermite adaptativa multidimensional, por defecto es 11, se recomienda no dar un valor muy grande a este parametro, por que afectara de manera significativamente los tiempos de convergencia del modelo.}

\item{pruning}{Es un valor booleano que indica si se utiliza pruning o no, para la cuadratura de Gauss-Hermite adaptativa multidimensional, por defecto es TRUE.}
}
\description{
function RMM.ZOIP fits a mixed ZOIP regression model with random intercepts
normals in the mean and dispersion parameter, the estimation is done via maximum likelihood
and the gauss-hermite adaptive quadrangle with or without pruning. the model may or may not include effects
fixed in any of its parameters, just as it can be a bilaterally inflated model, unilaterally or without parameters
inflated.
}
\examples{

library(ZOIP)
N<-21

Times <- c(2, 10, 20, 40) # cantidad de dias

subject <- rep(1:N, each=length(Times)) # numero de sujetos en la muestra repetidos tantas veces haya dias

Days <- rep(Times, times=N)
b0i <- rep(rnorm(n=N,sd=1), each=length(Times))
b1i <- rep(rnorm(n=N,sd=0.5), each=length(Times))

neta <- (1.6+b0i)-1.3*log(Days)
neta2<-(0.1+b1i)-0.8*log(Days)


mu <- 1 / (1 + exp(-neta))
sigma <- 1 / (1 + exp(-neta2))

p0 <- 0.1
p1 <- 0.1

mu[mu==1] <- 0.999
mu[mu==0] <- 0.001

sigma[sigma==1] <- 0.999
sigma[sigma==0] <- 0.001


family<-'R-S'

Y <- rZOIP(n=length(mu), mu = mu, sigma = sigma ,p0=p0,p1=p1,family=family)
base<-data.frame(Y,Days,subject)


formula.mu <- Y ~ log(Days)
formula.sigma<-~log(Days)
formula.p0<-~1
formula.p1<-~1

formula.random=~1 | subject

link<-c('logit','logit','identity','identity')


optimizer<-'nlminb'
n.points <-11
pruning <- TRUE

mod<-RMM.ZOIP(formula.mu=formula.mu,formula.sigma=formula.sigma,formula.p0=formula.p0,formula.p1=formula.p1,data=base,
              formula.random=formula.random,link=link,family=family,optimizer=optimizer,
              n.points=n.points,pruning=pruning)
mod
}
